Added code but still not working

+++ CaloriesCatcher.UI/Model/UserDetail.cs
@@ -6,7 +6,7 @@
         public string UserId { get; set; }
         public int Height { get; set; }
         public int Weight { get; set; }
-        public int DailyCalories { get; set; }
+        public Decimal DailyCalories { get; set; }
         public DateTime BirthDate { get; set; }
         public string Gender { get; set; } = string.Empty;
         public int CalorieId { get; set; }
